config/pyproject.toml
[project]
name = "deepseek-engineer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.58.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]

[tool.coverage]
run = { source = ["src"] }
report = { show_missing = true, skip_covered = true }
